Fix Linux compile for HVF bootargs helper
